+++ ci/tests/reload/realod.go
@@ -1,0 +1,27 @@
+// Package reload is the test suits for reload config files.
+package reload
+
+import (
+	"io"
+	"os"
+)
+
+func cp(src, dst string) error {
+	in, err := os.Open(src)
+	if err != nil {
+		return err
+	}
+	defer in.Close()
+
+	out, err := os.Create(dst)
+	if err != nil {
+		return err
+	}
+	defer out.Close()
+
+	_, err = io.Copy(out, in)
+	if err != nil {
+		return err
+	}
+	return nil
+}