+++ pkg/hashkit/ketama.go
@@ -54,15 +54,20 @@ func newRingWithHash(hash func([]byte) uint) (h *HashRing) {
 	return
 }
 
-// Init init hash ring with nodes.
+// Init init ring.
 func (h *HashRing) Init(nodes []string, spots []int) {
+	h.lock.Lock()
+	h.init(nodes, spots)
+	h.lock.Unlock()
+}
+
+// Init init hash ring with nodes.
+func (h *HashRing) init(nodes []string, spots []int) {
 	if len(nodes) != len(spots) {
 		panic("nodes length not equal spots length")
 	}
-	h.lock.Lock()
 	h.nodes = nodes
 	h.spots = spots
-	h.lock.Unlock()
 	var (
 		ticks          []nodeHash
 		svrn           = len(nodes)
@@ -117,6 +122,7 @@ func (h *HashRing) AddNode(node string, spot int) {
 		exitst  bool
 	)
 	h.lock.Lock()
+	defer h.lock.Unlock()
 	for i, nd := range h.nodes {
 		tmpNode = append(tmpNode, nd)
 		if nd == node {
@@ -127,13 +133,12 @@ func (h *HashRing) AddNode(node string, spot int) {
 			tmpSpot = append(tmpSpot, h.spots[i])
 		}
 	}
-	h.lock.Unlock()
 	if !exitst {
 		tmpNode = append(tmpNode, node)
 		tmpSpot = append(tmpSpot, spot)
 		log.Infof("add node %s spot %d", node, spot)
 	}
-	h.Init(tmpNode, tmpSpot)
+	h.init(tmpNode, tmpSpot)
 }
 
 // DelNode delete a node from the hash ring.
@@ -146,6 +151,7 @@ func (h *HashRing) DelNode(n string) {
 		del     bool
 	)
 	h.lock.Lock()
+	defer h.lock.Unlock()
 	for i, nd := range h.nodes {
 		if nd != n {
 			tmpNode = append(tmpNode, nd)
@@ -155,9 +161,8 @@ func (h *HashRing) DelNode(n string) {
 			log.Info("ketama del node ", n)
 		}
 	}
-	h.lock.Unlock()
 	if del {
-		h.Init(tmpNode, tmpSpot)
+		h.init(tmpNode, tmpSpot)
 	}
 }